Update lp_HairKey.py
Fixed an issue where working in log messed with the ouput in two different ways.
Added a "saturation"-option for the applied effect. 

Changed the inner workings; choice of screen-colour is not made automatically by selected colour, therefore a new Pulldown menu is used. No integrated Constant as cleanplate anymore, so if a Constant colour is wanted it needs to be connected manually into the mask input. Also added a pulldown choice for the despill.
I made the choice to ban out the automatition as I find it more fancy than useful, and I don't want to embrace the usage of constants. Also it was not very userfriendly for people who dislike tooltips :)