Update lp_ColourSmear.py
Now based on the new EdgeExtend Node, I needed to rework the tool a bit. The raw results should be more or less the same, but it now works a bit differently. Some features went offline because of the workflow, but speed and quality should be greatly improved.
Shout out to Frédéric Devernay who not only introduced EdgeExtend, bit also bothering with me at all :)