update and renaming
I have no idea how to rename everything :( I overhauled all my despill-nodes (Nuke, Fusion, Natron) and renamed them, to get it a bit more uniform. New name is lp_ChillSpill

I reworked all the math and algorithms to get rid of all the SeExpr-nodes, as they were unusably slow. Now, everything is done with Shuffles and Merges, making it much faster. The tree is now a mess (may I add a feature-suggestion here? A "ChannelMath" node, something like a Merge, would be awesome; a Node that lets me choose a channel from each input, and a channel where to shuffle the result too, and an operation; like ChannelMerge in Nuke, so helpful and helps with some mess), but it works.
I also added a new function, "channel amp"; it amps the replacement channels BEFORE the algorithm, this can help with difficult despills. I also rebuild the tool that the average weight is not bypassable anymore, as the decision of this was only the decreased speed. Now it's not needed anymore :)