Create lp_Feather_adv.py
Feather_adv(anced) is a nifty little tool that does not more than what its name implies: feathering an alpha. In my professional life, this tool is so important for my workflow that I key-binded it to Alt+A.
However it can do more than only eroding and blurring it, it can also only feather inward/outward like you expect from a Roto-node. But with any shape :)
 It features different falloff-operations inspired by Nuke, but also a manual one (like in the Roto).